Here be dragons
There was a problem where (due to numerical issues) after pushing into a
particular feature (and disabling ignored epsilons), other
previously-ignored choices / branches became possible to select.  This
produced features whose choice list did not match the order of token ids in
the tape, which triggered an assertion that I left as a sanity check.

The fix is to store the list of choices as a set rather than a list, so
it is always sorted.  Because token ids in the tape are monotonically
increasing, the set of choices is guaranteed to happen in the same order
as the token tape, making everyone happier.

Resolves issue #37 (probably).