Add blend-clearance-inset
This is a function I like, to insert a shape into the surface of another one, with a smooth bevel surrounding it, as though pressing a marble partway into soft clay.

I'm admittedly unsure about whether this is getting too specific for a general CSG library, or whether you'd like to offer as much to the user as possible? (I'm also not certain whether its usage would be obvious just from reading it, especially with this name. Maybe "dimple inset"?)