Fix for blend-difference
Just found a subtle but important oversight in how I'd written blend-difference; it worked fine if used alone, but if nested, subsequent ones would be added instead of subtracted because the signs were reversed. This should be right now, sorry about the mistake; a function like this is a bit abstract to be sure about at a glance. :)

(I also got rid of the extraneous zero while I was in there.)