Less redundant definitions for the twirl functions
This shouldn't change anything on the front end, and was mostly done as an exercise for me, but it should hopefully lay out a useful way of generalizing axis-dependent transformation functions for anyone writing them in the future!

In fact, if I can get the hang of how to pass the remaining arguments through to another function in a generic way (or to pass them as a tuple or something), I bet the `generic-twirl-n` function could be generalized enough to be used for any function like this -- or maybe even to be passed to it as an argument, as in `(rotate-axis (sphere 1) transform-axis-y 2)` rather than `(rotate-y (sphere 1) 2)`? And similarly for the `center` argument... I'm not sure whether that would be simplifying things or complicating them at this point, so it's just an idea for now -- I like flexibility and reducing redundancy, but that might be starting to make the syntax for the end user too messy... Anyway, food for thought!