Remove a particularly shaky test case
The given test case was shaky due to sqrt, and was failing because of a
chain events down past the 10th digit of precision.  Even without
neighbor sharing, if I change the structure of circle (to remove the
sqrt), it started failing.

The new test is less aggressive, though I'm not totally sure that it's
testing the exact same thing...