Relicense libfive core under MPL.
This allows folks to statically link it, while still requiring changes
to the core library to be committed back to the community.  This is
compatible with my goals for the project: I don't have a strong opinion
on static vs dynamic linking, and certain platforms (e.g. iOS) make it
very hard to comply with the LGPL.

Looking through the list of contributors on Github, the only folks
(besides me) that have made nontrivial changes to the core library
are part of the organization that requested this change, so we can
make this license change without running into copyright issues.