Working on a robust solution to edge-pair problem
The edge-pair problem is actually an issue with how unique indexes are
assigned to corners, which are the only subspace that can be shared
between cells of different sizes.  We end up with the same corner having
different indexes, because cells of different sizes don't correctly
share it.

I'm planning to fix this by special-casing corners within assignIndices,
rather than changing anything about neighbor lookups (which I tried
earlier).