Change how we do bounded QEF solving
Previously, we'd pick the highest-dimensioned QEF solution that was
within the bounding box.  However, that doesn't always work - the case
that I was seeing was where one of the corners was the correct solution,
but that corner was considered out-of-bounds for the 1D check, so the
system selected a less optimal point on an edge instead.