FIX: Add file watcher on directory
Some programs like vim save a temporary file and move it to the new
location.
The QFileSystemWatcher will receive the notification "deleted" on the
current file. It does not notice that a new file is being created.
This patch will also listen to the underlying directory which will also
trigger a reload when a new file is being added. This should make the
file watch function much more robust.i